共通アイテムコンポーネントを追加し、結果行と提案アイテムの新しい実装を作成。不要な履歴アイテムと提案アイテムのファイルを削除。
