+++ src/components/common/result/item/CommonItem.tsx
@@ -1,4 +1,7 @@
+import React from "react";
+
 interface ResultItemProps {
+  key: React.Key;
   className?: string | undefined;
   onClick?: (event: React.MouseEvent) => void;
   children: React.ReactNode;
@@ -8,6 +11,7 @@ interface ResultItemProps {
 }
 
 export default function CommonItem({
+  key,
   className,
   onClick,
   leftContent,
@@ -15,18 +19,20 @@ export default function CommonItem({
   children,
 }: ResultItemProps) {
   return (
-    <div
-      className={`
+    <li key={key}>
+      <div
+        className={`
         flex items-center justify-between w-full px-4 py-2 text-left rounded-lg
         ${className}
       `}
-      onClick={onClick}
-    >
-      <div className="flex items-center min-w-full space-x-2">
-        {leftContent && <div className="flex-none">{leftContent}</div>}
-        <div className="flex-1">{children}</div>
-        {rightContent && <div className="flex-none">{rightContent}</div>}
+        onClick={onClick}
+      >
+        <div className="flex items-center min-w-full space-x-2">
+          {leftContent && <div className="flex-none">{leftContent}</div>}
+          <div className="flex-1">{children}</div>
+          {rightContent && <div className="flex-none">{rightContent}</div>}
+        </div>
       </div>
-    </div>
+    </li>
   );
 }
